+++ hub/initialize.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/greenplum-db/gpupgrade/greenplum/connection"
 	"github.com/greenplum-db/gpupgrade/idl"
 	"github.com/greenplum-db/gpupgrade/step"
 	"github.com/greenplum-db/gpupgrade/upgrade"
@@ -32,7 +33,17 @@ func (s *Server) Initialize(req *idl.InitializeRequest, stream idl.CliToHub_Init
 	}()
 
 	st.Run(idl.Substep_saving_source_cluster_config, func(stream step.OutStreams) error {
-		err = s.Config.AddClusters(req.GetPorts())
+		db, err := connection.Bootstrap(idl.ClusterDestination_source, s.Source.GPHome, s.Source.CoordinatorPort())
+		if err != nil {
+			return err
+		}
+		defer func() {
+			if cErr := db.Close(); cErr != nil {
+				err = errorlist.Append(err, cErr)
+			}
+		}()
+
+		err = s.Config.AddClusters(db, req.GetPorts())
 		if err != nil {
 			return err
 		}
